improving performance of EKCoreDataImporter
Using NSMutableSet and addObject, instead of NSSet and setByAddingObject for
accumulating existing primary keys within the import. NSSet:setByAddingObject
was creating copy of the set everytime a new primary keys was added which made
the process slow down exponentially with growing amount of primary keys.

Running the EKCoreDataImportBenchmarkSuite with howManyTimes set to 5000 used to
take around 30 seconds before the change, and after dropped just over 1 second.